Refactor QR code generation: move to_qr_code function to qrcode_utils and update response field names
